// src/app/apps/MP3Player/interfaces/Music.ts
export interface Music {
  album_id: string;
  album_image: string;
  album_name: string;
  artist_id: string;
  artist_idstr: string;
  artist_name: string;
  audio: string;
  audiodownload: string;
  duration: number;
  id: string;
  image: string;
  license_ccurl: string;
  name: string;
  position: number;
  prourl: string;
  readableDuration: string;
  releasedate: string;
  shareurl: string;
  shorturl: string;
}